+++ lib/core/index.d.ts
@@ -2,6 +2,6 @@ export { HttpException } from './exceptions/http-exception';
 export { MiddlewareBuilder } from './middlewares/builder';
 export { ModuleRef } from './injector/module-ref';
 export * from './services/reflector.service';
-export * from './nest-factory';
+export { NestFactory } from './nest-factory';
 export * from './nest-application';
 export * from './nest-application-context';