+++ src/core/application-config.ts
@@ -33,16 +33,24 @@ export class ApplicationConfig implements ConfigurationProvider {
     return this.ioAdapter;
   }
 
+  public addGlobalPipe(pipe: PipeTransform<any>) {
+    this.globalPipes.push(pipe);
+  }
+  
   public useGlobalPipes(...pipes: PipeTransform<any>[]) {
-    this.globalPipes = pipes;
+    this.globalPipes = this.globalPipes.concat(pipes);
   }
 
   public getGlobalFilters(): ExceptionFilter[] {
     return this.globalFilters;
   }
 
+  public addGlobalFilter(filter: ExceptionFilter) {
+    this.globalFilters.push(filter);
+  }
+
   public useGlobalFilters(...filters: ExceptionFilter[]) {
-    this.globalFilters = filters;
+    this.globalFilters = this.globalFilters.concat(filters);
   }
 
   public getGlobalPipes(): PipeTransform<any>[] {
@@ -53,15 +61,23 @@ export class ApplicationConfig implements ConfigurationProvider {
     return this.globalInterceptors;
   }
 
+  public addGlobalInterceptor(interceptor: NestInterceptor) {
+    this.globalInterceptors.push(interceptor);
+  }
+
   public useGlobalInterceptors(...interceptors: NestInterceptor[]) {
-    this.globalInterceptors = interceptors;
+    this.globalInterceptors = this.globalInterceptors.concat(interceptors);
   }
 
   public getGlobalGuards(): CanActivate[] {
     return this.globalGuards;
   }
 
+  public addGlobalGuard(guard: CanActivate) {
+    this.globalGuards.push(guard);
+  }
+
   public useGlobalGuards(...guards: CanActivate[]) {
-    this.globalGuards = guards;
+    this.globalGuards = this.globalGuards.concat(guards);
   }
 }