+++ src/core/nest-factory.ts
@@ -5,6 +5,7 @@ import { NestContainer } from './injector/container';
 import { ExceptionsZone } from './errors/exceptions-zone';
 import { NestModuleMetatype } from '@nestjs/common/interfaces/modules/module-metatype.interface';
 import { Logger } from '@nestjs/common/services/logger.service';
+import { NestApplicationOptions } from '@nestjs/common/interfaces/nest-application-options.interface';
 import { messages } from './constants';
 import { NestApplication } from './nest-application';
 import { isFunction } from '@nestjs/common/utils/shared.utils';
@@ -19,34 +20,41 @@ import { MetadataScanner } from './metadata-scanner';
 import { MicroservicesPackageNotFoundException } from './errors/exceptions/microservices-package-not-found.exception';
 import { NestApplicationContext } from './nest-application-context';
 import { HttpsOptions } from '@nestjs/common/interfaces/https-options.interface';
+import { NestApplicationContextOptions } from '@nestjs/common/interfaces/nest-application-context-options.interface';
+import { NestMicroserviceOptions } from '@nestjs/common/interfaces/microservices/nest-microservice-options.interface';
+import { ApplicationConfig } from './application-config';
 
 const { NestMicroservice } =
   optional('@nestjs/microservices/nest-microservice') || ({} as any);
 
 export class NestFactoryStatic {
-  private container = new NestContainer();
-  private instanceLoader = new InstanceLoader(this.container);
-  private logger = new Logger('NestFactory', true);
-  private dependenciesScanner = new DependenciesScanner(
-    this.container,
-    new MetadataScanner(),
-  );
-
+  private readonly logger = new Logger('NestFactory', true);
   /**
    * Creates an instance of the NestApplication (returns Promise)
-   *
-   * @param  {} module Entry (root) application module class
-   * @param  {} express Optional express() server instance
    * @returns an `Promise` of the INestApplication instance
    */
   public async create(
-    module,
-    express = ExpressAdapter.create(),
-    httpsOptions: HttpsOptions = null, 
+    module: any,
+    expressOrOptions?: any | NestApplicationOptions,
+    options?: NestApplicationOptions,
   ): Promise<INestApplication> {
-    await this.initialize(module, express);
+    const isExpressInstance = expressOrOptions && expressOrOptions.response;
+    const [expressInstance, appOptions] = isExpressInstance
+      ? [expressOrOptions, options]
+      : [ExpressAdapter.create(), expressOrOptions];
+
+    const container = new NestContainer();
+    const applicationConfig = new ApplicationConfig();
+
+    this.applyLogger(appOptions);
+    await this.initialize(module, container, applicationConfig, expressInstance);
     return this.createNestInstance<NestApplication>(
-      new NestApplication(this.container, express, httpsOptions),
+      new NestApplication(
+        container,
+        expressInstance,
+        applicationConfig,
+        appOptions,
+      ),
     );
   }
 
@@ -54,20 +62,23 @@ export class NestFactoryStatic {
    * Creates an instance of the NestMicroservice (returns Promise)
    *
    * @param  {} module Entry (root) application module class
-   * @param  {MicroserviceConfiguration} config Optional microservice configuration
+   * @param  {NestMicroserviceOptions} options Optional microservice configuration
    * @returns an `Promise` of the INestMicroservice instance
    */
   public async createMicroservice(
     module,
-    config?: MicroserviceConfiguration,
+    options?: NestMicroserviceOptions,
   ): Promise<INestMicroservice> {
     if (!NestMicroservice) {
       throw new MicroservicesPackageNotFoundException();
     }
+    const container = new NestContainer();
+    const applicationConfig = new ApplicationConfig();
 
-    await this.initialize(module);
+    this.applyLogger(options);
+    await this.initialize(module, container, applicationConfig);
     return this.createNestInstance<INestMicroservice>(
-      new NestMicroservice(this.container, config as any),
+      new NestMicroservice(container, options as any, applicationConfig),
     );
   }
 
@@ -75,17 +86,22 @@ export class NestFactoryStatic {
    * Creates an instance of the NestApplicationContext (returns Promise)
    *
    * @param  {} module Entry (root) application module class
+   * @param  {NestApplicationContextOptions} options Optional Nest application configuration
    * @returns an `Promise` of the INestApplicationContext instance
    */
   public async createApplicationContext(
     module,
+    options?: NestApplicationContextOptions,
   ): Promise<INestApplicationContext> {
-    await this.initialize(module);
+    const container = new NestContainer();
 
-    const modules = this.container.getModules().values();
+    this.applyLogger(options);
+    await this.initialize(module, container);
+
+    const modules = container.getModules().values();
     const root = modules.next().value;
     return this.createNestInstance<INestApplicationContext>(
-      new NestApplicationContext(this.container, [], root),
+      new NestApplicationContext(container, [], root),
     );
   }
 
@@ -93,13 +109,25 @@ export class NestFactoryStatic {
     return this.createProxy(instance);
   }
 
-  private async initialize(module, express = null) {
-    this.container.setApplicationRef(express);
+  private async initialize(
+    module,
+    container: NestContainer,
+    config = new ApplicationConfig(),
+    express = null,
+  ) {
+    const instanceLoader = new InstanceLoader(container);
+    const dependenciesScanner = new DependenciesScanner(
+      container,
+      new MetadataScanner(),
+      config,
+    );
+    container.setApplicationRef(express);
     try {
       this.logger.log(messages.APPLICATION_START);
       await ExceptionsZone.asyncRun(async () => {
-        this.dependenciesScanner.scan(module);
-        await this.instanceLoader.createInstancesOfDependencies();
+        dependenciesScanner.scan(module);
+        await instanceLoader.createInstancesOfDependencies();
+        dependenciesScanner.applyApplicationProviders();
       });
     } catch (e) {
       process.abort();
@@ -129,6 +157,13 @@ export class NestFactoryStatic {
       }
       return receiver[prop];
     };
+  }
+
+  private applyLogger(options: NestApplicationContextOptions | undefined) {
+    if (!options || !options.logger) {
+      return;
+    }
+    Logger.overrideLogger(options.logger);
   }
 }